Rename preprocess to indexing and updated documentation and makefile
