+++ .gitignore
@@ -12,7 +12,7 @@ data/clean/filtered_data_100k.csv
 data/embeddings/embeddings_100k.npz
 /data/raw
 *.npz
-preprocess/images_100k_v2.zip
+indexing/images_100k_v2.zip
 data/csv/sample_100k_v2.csv
 .env
 **/__pycache__/
