+++ src/indexing/__init__.py
@@ -1,0 +1,3 @@
+"""
+Indexing module for handling data loading, embedding generation, and FAISS index computation.
+"""
