+++ src/indexing/compute_faiss_index.py
