--- src/preprocess/__init__.py
@@ -1,3 +1,0 @@
-"""
-Preprocessing module for handling data loading, embedding generation, and FAISS index computation.
-"""
